Refactored edit profile modal and edit bio modal into one modal
refs #227
