+++ app/components/user-edit-profile-modal.js
@@ -3,10 +3,19 @@ import { computed }  from '@ember/object';
 import Changeset from 'ember-changeset';
 import lookupValidator from 'ember-changeset-validations';
 import UserValidations from '../validations/user';
+import { filterBy, lt } from '@ember/object/computed';
+import { next }  from '@ember/runloop';
+import { inject as service } from '@ember/service';
+
+const MAX_AUTHORS = 5;
+const MAX_BOOKS = 5;
+
 const DEFAULT_TAB = 'overview';
 
 export default Component.extend({
   tagName: '',
+
+  store: service(),
 
   tab: null,
   open: null,
@@ -17,7 +26,7 @@ export default Component.extend({
   stat3: null,
   changeset: null,
   
-  
+  //* init *//
   init(){
     this._super(...arguments);
     //define an object to store stats properties 
@@ -49,6 +58,12 @@ export default Component.extend({
     let cs = new Changeset(this.get('user'), lookupValidator(UserValidations), UserValidations);
     this.set('changeset', cs);
   },
+  
+  /* computed properties */
+  canAddAnotherAuthor: lt('favoriteAuthors.length', MAX_AUTHORS),
+  canAddAnotherBook: lt('favoriteBooks.length', MAX_BOOKS),
+  favoriteAuthors: filterBy('user.favoriteAuthors', 'isDeleted', false),
+  favoriteBooks: filterBy('user.favoriteBooks', 'isDeleted', false),
 
   statsWordCountSelected: computed('stats1','stats2', 'stats3', function(){
     let key = "statsWordCountEnabled";
@@ -112,6 +127,8 @@ export default Component.extend({
     }
   }),
 
+  
+  /* actions */
   actions: {
     onHidden() {
       this.get('user').rollbackExternalLinks();
@@ -154,8 +171,55 @@ export default Component.extend({
       if (s3) {
         cs.set(s3, true);
       }
+    },
+    
+    addAuthor() {
+      if (this.get('canAddAnotherAuthor')) {
+        this._addAuthor();
+      }
+    },
+
+    addBook() {
+      if (this.get('canAddAnotherBook')) {
+        this._addBook();
+      }
+    },
+
+    deleteAuthor(author) {
+      author.deleteRecord();
+    },
+
+    deleteBook(book) {
+      book.deleteRecord();
+    },
+    
+    onHidden() {
+      this.get('user').rollbackFavorites();
+      let callback = this.get('onHidden');
+      if (callback) {
+        callback();
+      } else {
+        this.set('open', null);
+      }
     }
-  }
+    
+  },
+  /* component methods */
+  _addAuthor() {
+    let author = this.get('store').createRecord('favorite-author');
+    next(() => {
+      author.set('user', this.get('user'));
+    })
+    return author;
+  },
+
+  _addBook() {
+    let book = this.get('store').createRecord('favorite-book');
+    next(() => {
+      book.set('user', this.get('user'));
+    })
+    return book;
+  },
 
 });
