+++ app/templates/authenticated/users/show/index.hbs
@@ -58,13 +58,6 @@
       <div {{action "showEditBio"}} class="flip-corner-link" data-test-user-edit-bio>
         <img src="/images/global/flip_corner.png" alt="Edit User Bio" />
       </div>
-
-      {{user-edit-bio-modal
-        onHidden=(action "afterModalClose")
-        open=editBio
-        user=user
-        tab=editBioTab
-      }}
     {{/if}}
   </div>
   {{#if user.primaryProject.latestProjectChallenge }}
